Travis CI: Add pytest --doctest-modules maths
