+++ maths/abs_min.py
@@ -1,10 +1,11 @@
-from abs import abs_val
+from .abs import abs_val
+
 
 def absMin(x):
     """
-    # >>>absMin([0,5,1,11])
+    >>> absMin([0,5,1,11])
     0
-    # >>absMin([3,-10,-2])
+    >>> absMin([3,-10,-2])
     -2
     """
     j = x[0]
@@ -13,9 +14,11 @@ def absMin(x):
             j = i
     return j
 
+
 def main():
     a = [-3,-1,2,-11]
     print(absMin(a))  # = -1
+
 
 if __name__ == '__main__':
     main()
