+++ maths/lucas_series.py
@@ -1,0 +1,14 @@
+# Lucas Sequence Using Recursion
+
+def recur_luc(n):
+  if n == 1:
+    return n
+  if n == 0:
+    return 2
+  return (recur_luc(n-1) + recur_luc(n-2))
+
+if __name__ == "__main__":
+  limit = int(input("How many terms to include in Lucas series:").strip())
+  print("Lucas series:")
+  for i in range(limit):
+    print(recur_luc(i))
